requirements.txt
# Make sure clams-python version is explicitly specified, at least the lower bound
clams-python==1.1.3

easyocr==1.7.0
chardet==5.2.0
torch~=2.0.1
Pillow==9.5.0
numpy~=1.24.4
transformers==4.38.2

timm==0.9.16

